backend/requirements.txt
Flask>=2.0.0
Pillow>=8.0.0
numpy>=1.20.0
flask-cors>=3.0.0
Werkzeug>=2.0.0